Refactor code. Add DOI admin menu item.
