+++ public/static/locales/en/common.json
@@ -15,6 +15,7 @@
     "deTitle": "Discovery Environment",
     "discard": "Discard",
     "discovery": "<b>D</b>iscovery",
+    "doi": "DOI",
     "done": "Done",
     "drawerMenuAriaLabel": "drawer menu",
     "edit": "Edit",
