+++ src/components/layout/styles.js
@@ -1,0 +1,142 @@
+const drawerWidth = 235;
+const miniDrawerWidth = 65;
+export default (theme) => ({
+    root: {
+        flex: 1,
+    },
+    appBar: {
+        backgroundColor: theme.palette.primary,
+        zIndex: theme.zIndex.drawer + 1,
+        width: "100%",
+        [theme.breakpoints.up("sm")]: {
+            marginLeft: miniDrawerWidth,
+            transition: theme.transitions.create(["width", "margin"], {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.leavingScreen,
+            }),
+            width: `calc(100% - ${miniDrawerWidth + 1}px)`,
+        },
+    },
+    appBarShift: {
+        [theme.breakpoints.up("sm")]: {
+            marginLeft: drawerWidth,
+            width: `calc(100% - ${drawerWidth}px)`,
+            transition: theme.transitions.create(["width", "margin"], {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.enteringScreen,
+            }),
+        },
+    },
+    drawerIcon: {
+        flexShrink: 0,
+        marginRight: theme.spacing(3.5),
+        width: 32,
+    },
+    userIcon: {
+        cursor: "pointer",
+        backgroundColor: theme.palette.success.main,
+        color: theme.palette.success.contrastText,
+        width: theme.spacing(5),
+        height: theme.spacing(5),
+        marginLeft: theme.spacing(1),
+        "&:hover": {
+            textDecoration: "underline",
+            textDecorationColor: theme.palette.success.contrastText,
+        },
+    },
+    accountIcon: {
+        cursor: "pointer",
+        color: theme.palette.primary.contrastText,
+        backgroundColor: theme.palette.primary.main,
+        width: theme.spacing(5),
+        height: theme.spacing(5),
+    },
+    icon: {
+        color: theme.palette.info.contrastText,
+    },
+    appBarIcon: {
+        [theme.breakpoints.up("sm")]: {
+            margin: theme.spacing(2),
+        },
+        [theme.breakpoints.down("xs")]: {
+            margin: theme.spacing(0.2),
+        },
+    },
+    drawer: {
+        width: drawerWidth,
+        flexShrink: 0,
+        whiteSpace: "nowrap",
+    },
+    drawerOpen: {
+        backgroundColor: theme.palette.info.main,
+        color: theme.palette.info.contrastText,
+        width: drawerWidth,
+        [theme.breakpoints.up("sm")]: {
+            transition: theme.transitions.create("width", {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.enteringScreen,
+            }),
+        },
+    },
+    drawerClose: {
+        [theme.breakpoints.up("sm")]: {
+            backgroundColor: theme.palette.info.main,
+            color: theme.palette.info.contrastText,
+            transition: theme.transitions.create("width", {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.leavingScreen,
+            }),
+            overflowX: "hidden",
+            width: miniDrawerWidth,
+        },
+    },
+    content: {
+        width: "100%",
+        [theme.breakpoints.up("sm")]: {
+            width: `calc(100% - ${miniDrawerWidth + 1}px)`,
+            marginLeft: miniDrawerWidth,
+            padding: theme.spacing(1),
+        },
+    },
+    contentShift: {
+        [theme.breakpoints.up("sm")]: {
+            marginLeft: drawerWidth,
+            width: `calc(100% - ${drawerWidth}px)`,
+            transition: theme.transitions.create(["width", "margin"], {
+                easing: theme.transitions.easing.sharp,
+                duration: theme.transitions.duration.enteringScreen,
+            }),
+        },
+    },
+    profile: {
+        cursor: "pointer",
+        "&:hover": {
+            textDecoration: "underline",
+            textDecorationColor: theme.palette.info.contrastText,
+        },
+    },
+    listItem: {
+        cursor: "pointer",
+        "&:hover": {
+            textDecoration: "underline",
+            backgroundColor: theme.palette.primary.main,
+        },
+        color: theme.palette.info.contrastText,
+    },
+    listItemActive: {
+        cursor: "pointer",
+        textDecoration: "underline",
+        color: theme.palette.primary.contrastText,
+        backgroundColor: theme.palette.primary.main,
+    },
+    menuIcon: {
+        color: theme.palette.primary.contrastText,
+        [theme.breakpoints.up("sm")]: {
+            margin: theme.spacing(0.5),
+            color: theme.palette.info.contrastText,
+        },
+    },
+    nested: {
+        paddingLeft: theme.spacing(4),
+    },
+});
