+++ src/pages/admin/doi.js
@@ -1,0 +1,12 @@
+/**
+ *
+ * @author sriram
+ *
+ * A page for the DE Admin to access DOI requests.
+ *
+ */
+import React from "react";
+
+export default function Doi() {
+    return <div>Admin DOI Page! </div>
+}
